LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/css-module/import-css-module-basic-expected.txt

CONSOLE MESSAGE: line 1: TypeError: 'text/css' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: line 1: TypeError: 'text/css' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: line 1: TypeError: 'text/css' is not a valid JavaScript MIME type.


FAIL A CSS Module should load assert_equals: CSS module import should succeed expected "rgb(255, 0, 0)" but got "rgba(0, 0, 0, 0)"
FAIL An @import CSS Module should not load assert_equals: expected (string) "NotAllowedError" but got (undefined) undefined
PASS Malformed CSS should not load 
